split classes into separate files, tests 1-7 still pass
