+++ computer_player.rb
@@ -1,0 +1,7 @@
+class ComputerPlayer < Player
+  attr_reader :name
+
+  def initialize
+    @name = "HAL 9000"
+  end
+end
