+++ human_player.rb
@@ -1,0 +1,7 @@
+class HumanPlayer < Player
+  attr_reader :name
+
+  def initialize(name = "Dave")
+    @name = name
+  end
+end
