+++ player.rb
@@ -1,0 +1,2 @@
+class Player
+end
