+++ ship.rb
@@ -1,0 +1,7 @@
+class Ship
+  attr_reader :length
+
+  def initialize(length)
+    @length = length
+  end
+end
